security lvl 1 - projects and partners

+++ app/controllers/ProjectController.java
@@ -14,8 +14,10 @@ import play.data.Form;
 import play.db.jpa.Transactional;
 import play.mvc.Controller;
 import play.mvc.Result;
+import service.ActionsEnum;
 import service.Configuration;
 import service.ProjectConverter;
+import service.SecurityService;
 import views.data.MenuDto;
 import views.data.ProjectDto;
 import views.html.projects.projectDetail;
@@ -43,6 +45,10 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=true)
 	public static Result showAll(Integer page) {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_SHOW_ALL)) {
+			return redirect(routes.Application.accessDenied());
+		}
 		if (page == null) {
 			page = 0;
 			Logger.debug("No page number is given. Setting 0 (first page).");
@@ -52,17 +58,22 @@ public class ProjectController extends Controller{
 		Integer totalProjects = DAOs.getProjectDao().getNumberOfProjects();
 		Integer numberOfPages = totalProjects % Configuration.PAGE_SIZE == 0 ? totalProjects/Configuration.PAGE_SIZE : totalProjects/Configuration.PAGE_SIZE + 1; 
 		Logger.debug("Page with list of projects is shown. Number of projects id db is {}", totalProjects);
-		return ok(projects.render(ProjectConverter.convertListToDto(proj), getMainMenu(), numberOfPages, page));
+		return ok(projects.render(ProjectConverter.convertListToDto(proj), getMainMenu(user), numberOfPages, page));
 	}
 	
 	/**
 	 * Action shows form for adding new project
 	 * @return
 	 */
+	@Transactional(readOnly=true)
 	public static Result create() {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_CREATE)) {
+			return redirect(routes.Application.accessDenied());
+		}
 		Form<ProjectDto> projectForm = Form.form(ProjectDto.class);
 		Logger.debug("Page with form for creating new project is shown.");
-		return ok(projectsCreate.render(projectForm, getBackToListMenu()));
+		return ok(projectsCreate.render(projectForm, getBackToListMenu(user)));
 	}
 	/**
 	 * Action saves new project
@@ -70,6 +81,10 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result saveNewProject() {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_CREATE)) {
+			return redirect(routes.Application.accessDenied());
+		}
 		Form<ProjectDto> projectForm = Form.form(ProjectDto.class).bindFromRequest();
 		Project newProject = ProjectConverter.convertToEntity(projectForm.get());
 		newProject.setVisible(true);
@@ -84,13 +99,17 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result detail(Long projectId) {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_DETAIL)) {
+			return redirect(routes.Application.accessDenied());
+		}
 		Project project = DAOs.getProjectDao().findById(projectId);
 		if (project == null) {
 			Logger.info("Partner with id {} was not found, detail can not be shown.", projectId);
 			return redirect(routes.ProjectController.projectNotFound(projectId));
 		} else {
 			Logger.debug("Partner detail page is shown.");
-			return ok(projectDetail.render(ProjectConverter.convertToDto(project), getBackToListMenu(), null, new ArrayList<User>()));
+			return ok(projectDetail.render(ProjectConverter.convertToDto(project), getBackToListMenu(user), null, new ArrayList<User>()));
 		}
 	}
 	
@@ -101,6 +120,11 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result edit(Long projectId) {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_EDIT)) {
+			return redirect(routes.Application.accessDenied());
+		}		
+		//TODO tmichalicka - second level verification
 		ProjectDto dto = ProjectConverter.convertToDto(DAOs.getProjectDao().findById(projectId));
 		if (dto == null) {
 			Logger.info("Project with id {} was not found.", projectId);
@@ -108,7 +132,7 @@ public class ProjectController extends Controller{
 		}
 		Form<ProjectDto> projectForm = Form.form(ProjectDto.class).fill(dto);
 		Logger.debug("Page with form for editing project is shown. Edited project has id {} and name {}.", dto.getProjectId(), dto.getName());
-		return ok(projectsEdit.render(projectForm, getBackToListMenu(), false));
+		return ok(projectsEdit.render(projectForm, getBackToListMenu(user), false));
 	}
 	
 	/**
@@ -118,6 +142,10 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result delete(Long projectId) {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_DELETE)) {
+			return redirect(routes.Application.accessDenied());
+		}	
 		Project project = DAOs.getProjectDao().findById(projectId);
 		if (project != null) {
 			DAOs.getProjectDao().delete(project);
@@ -134,9 +162,10 @@ public class ProjectController extends Controller{
 	 * @param projectId
 	 * @return
 	 */
+	@Transactional(readOnly=true)
 	public static Result projectNotFound(Long projectId) {
 		Logger.debug("Project not found page is shown. Requested id was {}.", projectId);
-		return ok(projectNotFound.render(projectId, getBackToListMenu()));
+		return ok(projectNotFound.render(projectId, getBackToListMenu(SecurityService.fetchUser(session("authid")))));
 	}
 	
 	/**
@@ -146,6 +175,10 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result updateProject(Boolean forceNext) {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_EDIT)) {
+			return redirect(routes.Application.accessDenied());
+		}
 		Form<ProjectDto> projectForm = Form.form(ProjectDto.class).bindFromRequest();
 		Project project = ProjectConverter.convertToEntity(projectForm.get());
 		if (forceNext != null && forceNext) {
@@ -172,7 +205,7 @@ public class ProjectController extends Controller{
 			DAOs.getProjectDao().update(project);
 		} catch (OptimisticLockException e) {
 			Logger.info("Project {} was edited by another user. ", projectForm.get());
-			return ok(projectsEdit.render(projectForm, getBackToListMenu(), true));
+			return ok(projectsEdit.render(projectForm, getBackToListMenu(user), true));
 		}
 		Logger.debug("Project update operation was called.");
 		return redirect(routes.ProjectController.showAll(0));
@@ -181,15 +214,15 @@ public class ProjectController extends Controller{
 	 * Method returns list of items to left side menu. This implementation returns one item - back to list
 	 * @return
 	 */
-	private static List<MenuDto> getBackToListMenu() {
+	private static List<MenuDto> getBackToListMenu(User user) {
 		List<MenuDto> result = new ArrayList<MenuDto>();
-		
-		MenuDto back = new MenuDto();
-		back.setGlyphicon("triangle-left");
-		back.setLabel("Zpět na seznam projektů");
-		back.setUrl(routes.ProjectController.showAll(0).absoluteURL(request()));
-		result.add(back);
-		
+		if (SecurityService.hasAccess(user, ActionsEnum.PROJECT_SHOW_ALL)) {
+			MenuDto back = new MenuDto();
+			back.setGlyphicon("triangle-left");
+			back.setLabel("Zpět na seznam projektů");
+			back.setUrl(routes.ProjectController.showAll(0).absoluteURL(request()));
+			result.add(back);
+		}
 		return result;
 	}
 	
@@ -197,15 +230,15 @@ public class ProjectController extends Controller{
 	 * Method returns list of items to left side menu. This implementation returns one item - add new
 	 * @return
 	 */
-	private static List<MenuDto> getMainMenu() {
+	private static List<MenuDto> getMainMenu(User user) {
 		List<MenuDto> result = new ArrayList<MenuDto>();
-		
-		MenuDto newProject = new MenuDto();
-		newProject.setGlyphicon("plus");
-		newProject.setLabel("Přidat projekt");
-		newProject.setUrl(routes.ProjectController.create().absoluteURL(request()));
-		result.add(newProject);
-		
+		if (SecurityService.hasAccess(user, ActionsEnum.PROJECT_CREATE)) {
+			MenuDto newProject = new MenuDto();
+			newProject.setGlyphicon("plus");
+			newProject.setLabel("Přidat projekt");
+			newProject.setUrl(routes.ProjectController.create().absoluteURL(request()));
+			result.add(newProject);
+		}
 		return result;		
 	}
